Bug Fixes & Speed Improvements
- Fixed images not showing up
- Made seperate timer for drawing, so the game doesn't draw 100fps and instead draws 50fps
- Background Image Causes Lag, ned fix, very bvad

+++ Obstacle.cs
@@ -23,7 +23,7 @@ namespace _VP_Project___Crazy_Eater
             Size = size;
             Speed = speed;
             image = Image.FromFile("Images\\EnemyIMG.png");
-            Ratio = image.Width / image.Height;
+            Ratio =(float) image.Width / image.Height;
         }
         public void Draw(Graphics g)
         {
